# Make `extract_last_comment` 65x faster