--- a/lib/method_source/code_helpers.rb
+++ b/lib/method_source/code_helpers.rb
@@ -106,52 +106,20 @@ module MethodSource
     #
     # @param [Array<String>]  lines
     # @return [String]
-    def extract_last_comment_core(lines)
-      buffer = ""
-      found_at_least_one_comment = false
-
-      lines.each do |line|
-        # Add any line that is a valid ruby comment,
-        # but clear as soon as we hit a non comment line.
-        starts_with_comment = (line =~ /^\s*#/)
-        if starts_with_comment
-          found_at_least_one_comment = true
-        end
-
-        if starts_with_comment || (line =~ /^\s*$/)
-          # this works
-          buffer += line.lstrip
-          # this breaks with Ruby 2.7.2p137: execution doesn't
-          # continue past it and no exception is raised
-          # buffer << line.lstrip
-        else
-          # this often breaks
-          # buffer.replace("")
-          # this works
-          buffer = ""
-        end
-      end
-
-      return [ buffer, found_at_least_one_comment ]
-    end
-
     def extract_last_comment(lines)
-      found_at_least_one_comment = false
+      buffer = []
 
-      if lines.size > @@EXTRACT_LAST_COMMENT_THRESHOLD
-        # if the last comment is found over fewer lines then don't
-        # need to go over all lines
-        index_start = (lines.size - @@EXTRACT_LAST_COMMENT_THRESHOLD)
-        index_start = 0 if index_start < 0
-        index_stop = (lines.size-1)
-        buffer, found_at_least_one_comment = extract_last_comment_core(lines[index_start..index_stop])
+      lines.reverse.each do |line|
+        # Add any line that is a valid ruby comment, and stop as
+        # soon as we hit a non comment line.
+        if (line =~ /^\s*#/) || (line =~ /^\s*$/)
+          buffer.append(line.lstrip)
+        else
+          break
+        end
       end
 
-      if not found_at_least_one_comment
-        buffer, found_at_least_one_comment = extract_last_comment_core(lines)
-      end
-
-      buffer
+      buffer.reverse.join()
     end
 
     # An exception matcher that matches only subsets of SyntaxErrors that can be