In `extract_last_comment` don't go over all lines of the file if the last comment is in the last 100 lines.

When used with a benchmark that called `extract_last_comment` 56411 times:
- <del>this improved performance by 3.42x (reduced execution time from `20.15s` to `5.89s`)</del>
- this improved performance by 65.4x (reduced execution time from `20.15s` to `0.31s`) when parsing comments backwards.
- <del>when the median number of lines in the opened files was 264 lines.</del>

<del>I tried to tune this with a different threshold, like 50-80 lines or 150-500 lines, and ~100 lines worked well for the benchmark used.</del>